refactor: 테스트 공통메소드 클래스 추출
