+++ src/test/java/nextstep/subway/testutils/RestAssuredMethods.java
@@ -1,0 +1,40 @@
+package nextstep.subway.testutils;
+
+import io.restassured.RestAssured;
+import io.restassured.response.ExtractableResponse;
+import io.restassured.response.Response;
+import org.springframework.http.MediaType;
+
+public class RestAssuredMethods {
+    public static <T> ExtractableResponse<Response> post(String path, T requestBody) {
+        return RestAssured.given().log().all()
+                .body(requestBody)
+                .contentType(MediaType.APPLICATION_JSON_VALUE)
+                .when().post(path)
+                .then().log().all()
+                .extract();
+    }
+
+    public static <T> ExtractableResponse<Response> put(String path, T requestBody) {
+        return RestAssured.given().log().all()
+                .body(requestBody)
+                .contentType(MediaType.APPLICATION_JSON_VALUE)
+                .when().put(path)
+                .then().log().all()
+                .extract();
+    }
+
+    public static ExtractableResponse<Response> get(String path) {
+        return RestAssured.given().log().all()
+                .when().get(path)
+                .then().log().all()
+                .extract();
+    }
+
+    public static ExtractableResponse<Response> delete(String path) {
+        return RestAssured.given().log().all()
+                .when().delete(path)
+                .then().log().all()
+                .extract();
+    }
+}
